Val Framework compares dataset matches that do not include the temporal reference
This changes the behavior of the validation framework.

Before this a validation run with `(n, k)` where `k < n` only gave results for
combinations of datasets which included the temporal reference.

e.g. If 3 dataset (DS1, DS2 and DS3) where given and a metric calculator was run
with `(3, 2)` with the temporal reference DS1. Then only the combinations of
`DS1 - DS2` and `DS1 - DS3` gave results.

This rectifies this behavior to also return results for the combination `DS2 -
DS3` in the above example.

Tests still need to be adapted but I want to discuss this first to see if this
is what people expect.

There are also some edge cases. e.g. If 4 datasets are given but the
configuration of the metric calculator would be `(3, 2)`. I think we should
raise an error in these cases because that would mean that the temporal
reference is not included in all matches.