comparison of new and old matching implementation
This adds a test to check whether the new matching in
BasicTemporalMatching works as the old one in case there are no
duplicates in the reference frame.

It replaced the matcher in the other tests by a dummy that still uses
the old matching implementation.
The following commit will remove the dummy and set the new expected
values.