New CDF matching implementation (#259)
This is a major rewrite of the CDF matching implementation currently used in pytesmo. It implements a few bugfixes and performance improvements (see #255) and provides a new interface that separates the calculation of the percentile values from the scaling step.

It replaces the deprecated method scaling.cdf_match with a method that has a similar functionality and interface as scaling.cdf_beta_match, the currently recommended method (and removes thereby the old implementation of scaling.cdf_match).